Better trap logic for moving Dockerfiles around
Create a separate function that handles docker builds. This is only
required until Docker adds support for building with specific
Dockerfiles.

We setup a trap for ERR and RETURN signals. If the docker build fails,
ERR trap is triggered and Dockerfiles are restored. If it succeeds,
RETURN trap is triggered and the ERR trap is removed.