Make initialization more sophisticated
We now start mysqld after initialization, without network. We initialize
it, using CLI wherever possible instead of SQL files. After
initialization is done, we properly shutdown the server and then start
it again properly, using exec.

We also use 'source scl_source enable mysql55' instead of X_SCLS. It is
not supported as well, but should be a bit safer.