Remove ugly {user,group}add
The Fedora/CentOS/RHEL mariadb/mysql rpms add mysql user and group in
the same manner we do (it was copied from it). It is not supposed to be
changed in future releases. Adding the user and group ourselves is very
ugly and the plethora of flags is confusing. Instead of that, we'll just
test that the UID is 27 in the Dockerfile. Even if UID in upstream image
would change, we'll get a build failure.