Change remove_tmp_dockerfile
The `-z` flag returns `true` if a string length is 0. `[ ! -z "$var" ]` is equivalent to `[ -n "$var" ]` or `[ "$string" ]` and in either case is not testing what is necessary here - because the string in question, `"${DOCKERFILE_PATH}.version"`, will *never* be of length 0, this `if` statement is not doing much.