Add persistent volume claim to slave
This accommodates for slave restarts without needing to re-sync the
entire database, as well as having a hot backup of the data.

Changes to run-mysqld-slave are needed for proper pod start up with a
persistent volume.

Change default volume size to 1 GiB

This aligns the volume size with the offering in OpenShift Online and
mirrors the MySQL template defaults in
https://github.com/openshift/origin/blob/master/examples/db-templates/mongodb-persistent-template.json