[Updated] Replace '--initialize-insecure' by '--initialize'
When the container starts, there is a test to check if the data
directory ($MYSQL_DATADIR) is empty or not. If it is empty, we need to
initialize it.

MySQL provides to ways to accomplish this, depending on whether you want
the server to generate a random initial password for the
'root'@'localhost' account:

- '--initialize': A random initial password is generated.

- '--initialize-insecure': No root password is generated, the password
is empty.

This commit switches from the '--initialize-insecure' to '--initialize'
for the data directory setup.

One of the motivations behind it is to allow the utilization of the
'validate_password' plugin. Avoiding a chicken and egg problem, in which
this plugin doesn't allow the initial root password to be empty.

Alongside using the '--initialize' option, it works by extracting the
auto generated root password from the log file, and then using it to do
the initial user setup.

Differences from the first version:
- For security reasons a log file for the initialization step is no
longer created, the temporary password is 'grepped' directly from the
stdout
- If the MYSQL_ROOT_PASSWORD is not set, an empty password is set to the
root user