Remove calls of mysql_upgrade that is removed in 8.4 and NooP before anyway
Starting 8.0.16, mysql_upgrade became NooP and server executes all actions
automatically when necessary. After mysql_upgrade being labeled as
deprecated since this version, it is missing in 8.4.x entirely.

In addition to that, the file mysql_upgrade_info file has been deprecated
as well and removed in 8.4.x as well, which makes all the tests done
around upgrade and mysql_upgrade_info invalid.

Since we have a recent enough version in production for enough time,
it seems ok to remove the logic about upgrading the datadir entirely from
the container image and leave everything on the server itself.

The only datadir actions that seem to make sense even in 8.4.x are
analyze and optimize, so those are kept untouched.