[build resilience fix] Install DB packages via versioned names
MariaDB and MySQL package switched to the versioned package layout in Fedora and RHEL 10.
The packages can be available in various states, but in all cases where a specific version is required, specific version must be requested.

Thus container wanting MySQL 8.4 server specifically, must request to install 'mysql8.4-server' specifically,
instead of just 'mysql-server' which will give you any version that is selected to be the distribution default at that moment.

That creates an issue with 'rpm -V', as it does not accept virtual provides as package names for processing, failing.

The result of a discussion with the sclorg containers manitainers is that the 'rpm -V' is a remnant from the ages of 'yum', the 'dnf' predecessor
When 'yum' had to install a list of packages and some were not found, it happily installed only those found, not raising an error.
The 'rpm -V' here was a protection against that - a simple verification that all requested packages were actually found and installed.